Split SmartEnterFeature
