+++ tests/features/EnterOutdentIfLineIsEmptyFeature.test.ts
@@ -11,28 +11,6 @@ export default {
     // assert
     plugin.assertCurrentState(["- one", "\t- two", "\t- |"]);
   },
-  "enter should create newline on the same level": (plugin) => {
-    // arrange
-    plugin.applyState(["- one", "\t- two|"]);
-
-    // act
-    plugin.simulateKeydown("Enter");
-
-    // assert
-    plugin.assertCurrentState(["- one", "\t- two", "\t- |"]);
-  },
-  "enter should create newline on the child level if child exists": (
-    plugin
-  ) => {
-    // arrange
-    plugin.applyState(["- one", "\t- two|", "\t\t- three"]);
-
-    // act
-    plugin.simulateKeydown("Enter");
-
-    // assert
-    plugin.assertCurrentState(["- one", "\t- two", "\t\t- |", "\t\t- three"]);
-  },
   "enter should delete list item if it's last item and it's on the top level": (
     plugin
   ) => {
